(unfinished) CardDetailsDialog
Prepare the ground for CardDetailsDialog, with the operation in CardManager and deckAbstraction.

This is semi working right now-- the group name seems to not be working.
The abstraction layer should really hand the most straightforward data to the UI
so there might need to be some reassessment on how it's shipping the data over
and make it cleaner.